+++ src/main/scala/com/dataintuitive/viash/helpers/Exec.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.helpers
+package io.viash.helpers
 
 import sys.process.{Process, ProcessLogger}
 import java.io.{ByteArrayOutputStream, File, PrintWriter}