+++ src/main/scala/com/dataintuitive/viash/helpers/Scala.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.helpers
+package io.viash.helpers
 
 object Scala {
   implicit class AugmentedOption[T](opt: Option[T]) {