+++ src/main/scala/com/dataintuitive/viash/platforms/Platform.scala
@@ -15,15 +15,15 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.dataintuitive.viash.platforms
+package io.viash.platforms
 
-import com.dataintuitive.viash.functionality.Functionality
-import com.dataintuitive.viash.config.Config
-import com.dataintuitive.viash.helpers.IO
+import io.viash.functionality.Functionality
+import io.viash.config.Config
+import io.viash.helpers.IO
 import io.circe.yaml.parser
 import java.net.URI
 import requirements._
-import com.dataintuitive.viash.config.Version
+import io.viash.config.Version
 
 trait Platform {
   val `type`: String